Intermediate state on the Trait-based impl
I'm just going to commit this here for future reference before I'm
abandoning it in favor of a simple `enum` based implementation.

The reason for this is rather straight-forward: We use the trait to
unify over a common interface, i.e. `Element`. The problem, however, is
that we need at multiple points in the code the original type to
differentiate between edges and points. This can only be realized using
the `Any` trait, which allows the downcasting of an object to a concrete
type. I put some example code that details this here:
https://play.rust-lang.org/?version=nightly&mode=debug&edition=2021&gist=ba9a3c180e35bf2208705a189d3ad8cf
(can also be found at the bottom)

Now the thing is that we end up having a very fragile if-else chain
exhausting all (desirable) options for downcasting, as seen in the
playground, which is in the end even worse than hiding everything behind
an enum which can emulate the behavior of a trait just as well while
giving us always a more simple and faster option to extract the type
using a match pattern, which also is exhaustive and ensures correctness.
Also, we have to sacrifice the `Foo` trait (or at least I haven't found
a way to keep it in the time I tried).

Hence, I'm going to move on with an `enum` pattern.
You can read more here:
https://bennetthardwick.com/dont-use-boxed-trait-objects-for-struct-internals/

---

```rust
use std::any::Any;

// this is our trait. Seems simple enough
trait Foo: Any {
    fn some_fun(&self) -> bool;
}

// what follows are 2 dummy structs
struct Bar {
    pub field: String,
}

impl Foo for Bar {
    fn some_fun(&self) -> bool {
        false
    }
}

struct Baz {
    pub field: String,
}

impl Foo for Baz {
    fn some_fun(&self) -> bool {
        true
    }
}

// this is the function that is the problem: in it we need to clone the
element
// in question, requiring T: Clone
// approaches 1-3
/* fn test<T: Clone + Foo>(something: Box<T>) -> (Box<T>, Box<T>) {
    (something.clone(), something)
} */

// approach 4: This is the "solution" to the problem.
// The issue: This
fn test(something: &Box<dyn Any>) -> (Box<dyn Foo>, Box<dyn Foo>) {
    if let Some(v) = something.downcast_ref::<Bar>() {
        (Box::new(v.clone()), Box::new(v.clone()))
    } else if let Some(v) = something.downcast_ref::<Baz>() {
        (Box::new(v.clone()), Box::new(v.clone()))
    } else {
        panic!("failure")
    }
}

// needed for approach 3
trait NewTrait: Foo + Clone {}

fn main() {
    // approach 1: obviously doesn't work, because `Clone` is not
satisfied
    /* let v: Vec<Box<dyn Foo>> = vec![
        Box::new(Baz {
            field: "test".to_string(),
        }),
        Box::new(Bar {
            field: "test".to_string(),
        }),
    ]; */
    // approach 2: Let's require `Clone` then
    /* let v: Vec<Box<dyn Foo + Clone>> = vec![
        Box::new(Baz {
            field: "test".to_string(),
        }),
        Box::new(Bar {
            field: "test".to_string(),
        }),
    ]; */
    // approach 3: let's add the proposed supertrait
    /* let v: Vec<Box<dyn NewTrait>> = vec![
        Box::new(Baz {
            field: "test".to_string(),
        }),
        Box::new(Bar {
            field: "test".to_string(),
        }),
    ]; */

    // approach 4: looses us the `Foo` trait...
    let v: Vec<Box<dyn Any>> = vec![
        Box::new(Baz {
            field: "test".to_string(),
        }),
        Box::new(Bar {
            field: "test".to_string(),
        }),
    ];

    //approaches 1-3
    /* for elem in v {
        let _ = test(elem);
    } */

    // approach 4
    for elem in v {
        let _ = test(&elem);
    }
}
```