A lot of debugging, no success yet though
The problem that seems to persist is that the algorithm continuously
creates smaller and smaller triangles that then end up being so small
that some of the math starts to fail (because we get in the range of
natural floating-point drifts). I tried to limit the problem by not
marking triangles as bad anymore when their area is smaller than 0.1,
but that apparently isn't enough. I will continue with this tomorrow
morning after looking at preflow push.